+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## @i-vresse/wb-core 1.1.0 - 2022-08-10
+
+### Added
+
+* Molecule filename as index in array sub-forms ([#76](https://github.com/i-VRESSE/workflow-builder/issues/76))
+
+## @i-vresse/wb-form 1.1.0 - 2022-08-10
+
+### Added
+
+* Molecule filename as index in array sub-forms ([#76](https://github.com/i-VRESSE/workflow-builder/issues/76))
+
 ## 1.0.0 - 2022-08-02
 
 Initial release
@@ -13,7 +25,7 @@ Initial release
 <!-- 
 To record change to a package use a custom header like 
 
-## @i-vresse/wb-form@1.0.1 - 2022-08-02
+## @i-vresse/wb-form 1.0.1 - 2022-08-02
 
 ... Changes of form package ...